requirements.txt.txt
PyQt6
scapy
python-nmap
fpdf
requests
pyinstaller
